+++ CHANGELOG.md
@@ -5,11 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
 
-## [0.0.1] - (unreleased)
+### Changed
+
+### Removed
+
+
+## [0.0.1] - 2022-05-03
 ### Added
 - [CHANGELOG] Configuration for changelog.md [Patch]
 - [README] README file for this project [Patch]
 - [requirements] Added package requirements [Patch]
-- [base]-[core]-[image_processing]-[Image_Processing] Implemented extract_channel [Major]
-- [base]-[core]-[add_elements]-[add_layers] Added support to add black colour border on all 4 sides [Major]
+- [image-editor]-[core]-[image_processing]-[Image_Processing] Implemented extract_channel [Major]
+- [image-editor]-[core]-[add_elements]-[add_layers] Added support to add black colour border on all 4 sides [Major]